Create the basic test for report_ingest_status
